hjhdjd

+++ src/app/pages/user-list/user-list.component.ts
@@ -3,36 +3,45 @@ import { MatPaginator } from '@angular/material/paginator';
 import { MatTableDataSource } from '@angular/material/table';
 import { User } from 'src/app/core/models/user.model';
 import { UserListService } from 'src/app/services/user-list.service';
+import {MatDialog} from '@angular/material/dialog';
+import { DeleteDialogComponent } from 'src/app/dialogs/delete-dialog/delete-dialog.component';
+
 
 const COLUMNS_SCHEMA = [
   {
     key: 'id',
     type: 'number',
     label: 'Id',
-    disabled: true
+    disabled: true,
+    isAdd:false
   },
   {
     key: 'fullname',
     type: 'text',
     label: 'Full Name',
+    isAdd:false
   },
   {
     key: 'email',
     type: 'text',
     label: 'Email',
-    disabled: true
+    disabled: true,
+    isAdd:false
   },
   {
     key: 'password',
     type: 'text',
     label: 'Password',
+    isAdd:false
   },
   {
     key: 'isEdit',
     type: 'isEdit',
-    label: 'Action',
+    label: 'Action'
   },
 ];
+
+
 
 @Component({
   selector: 'app-user-list',
@@ -46,14 +55,16 @@ export class UserListComponent implements OnInit, AfterViewInit {
       fullname: '',
       email: '',
       password: '',
-      isEdit:false
+      isEdit:false,
+      isAdd:false
     },
   ];
   itemCount = 0;
   length = 0;
   // records: any[] = [];
   userCopy: User | null = null;
-  
+  nextId = 1;
+  newRow:any
 
   displayedColumns: string[] = COLUMNS_SCHEMA.map((col) => col.key);
   columnsSchema: any = COLUMNS_SCHEMA;
@@ -61,31 +72,41 @@ export class UserListComponent implements OnInit, AfterViewInit {
 
   @ViewChild(MatPaginator) paginator!: MatPaginator;
 
-  constructor(private _userService: UserListService) {}
+  constructor(private _userService: UserListService, public dialog: MatDialog) {}
 
   ngOnInit() {
     this.getUserList();
   }
 
-  getUserList():void {
-    this._userService.getUserData().subscribe((res) => {
-      this.userList = res;
-      this.userList = this.userList.map((user: User) => {
-        return { ...user, isEdit: false };
-      });
-      this.dataSource.data = this.userList;
-      this.itemCount = res.length;
-    });
-  }
+  // getUserList():void {
+  //   this._userService.getUserData().subscribe((res) => {
+  //     this.userList = res;
+  //     this.userList = this.userList.map((user: User) => {
+  //       return { ...user, isEdit: false };
+  //     });
+  //     this.dataSource.data = this.userList;
+  //     this.itemCount = res.length;
+  //   });
+  // }
 
   ngAfterViewInit() {
     this.dataSource.paginator = this.paginator;
   }
 
   deleteData(id: number) {
-    this._userService.deleteData(id).subscribe((record) => {
-      this.getUserList();
+    const dialogRef = this.dialog.open(DeleteDialogComponent, {
+      width: '450px', // Set the width of the dialog
+      data: { /* You can pass data to the dialog component if needed */ }
     });
+
+    dialogRef.afterClosed().subscribe(result => {
+      if (result === true) {
+        this._userService.deleteData(id).subscribe((record) => {
+          this.getUserList();
+        });
+      }
+    });
+    
   }
 
   editRow(row: User) {
@@ -99,7 +120,11 @@ export class UserListComponent implements OnInit, AfterViewInit {
       row.isEdit = false;
       localStorage.setItem('updatedUser', JSON.stringify(updatedUser));
       // console.log(localStorage);
-    });
+    },
+    (error) => {
+      console.error('An error occurred:', error);
+    }
+    );
   }
   cancelData(row:User){
     if (this.userCopy) {
@@ -108,4 +133,43 @@ export class UserListComponent implements OnInit, AfterViewInit {
       this.userCopy = null;
     }
   }
+  
+  addRow() {
+    
+    const newRow = this.prepareNewRow();
+    this.dataSource.data.push(newRow);
+    this.dataSource.data = [...this.dataSource.data]; // Trigger data source update
+  }
+  prepareNewRow(): User {
+    const newRow: User = {
+      id: this.nextId,
+      fullname: '',
+      email: '',
+      password: '',
+      isEdit: true,
+      isAdd:true
+    };
+    this.nextId++; // Increment the next available id for the next row
+    return newRow;
+  }
+  getMaxExistingId() {
+    const maxId = this.userList.reduce((max, user) => Math.max(max, user.id), 0);
+    this.nextId = maxId + 1;
+  }
+  getUserList(): void {
+    this._userService.getUserData().subscribe((res) => {
+      this.userList = res;
+      this.userList = this.userList.map((user: User) => {
+        return { ...user, isEdit: false };
+      });
+      this.dataSource.data = this.userList;
+      this.itemCount = res.length;
+      this.getMaxExistingId(); // Call it here to get the maximum existing id
+    });
+  }
+  saveNewRow(newRow:User){
+    newRow.isAdd = !newRow.isAdd;
+    console.log(newRow);
+  }
+  
 }
